+++ src/Vortice.D3DCompiler/Compiler.cs
@@ -1,7 +1,6 @@
-﻿// Copyright © Amer Koleci and Contributors.
+﻿// Copyright (c) Amer Koleci and Contributors.
 // Licensed under the MIT License (MIT). See LICENSE in the repository root for more information.
 
-using System;
 using Vortice.Direct3D;
 
 namespace Vortice.D3DCompiler;