+++ src/Vortice.Direct2D1/Effects/ConvolveMatrix.cs
@@ -1,8 +1,13 @@
-﻿// Copyright © Amer Koleci and Contributors.
+﻿// Copyright (c) Amer Koleci and contributors.
 // Licensed under the MIT License (MIT). See LICENSE in the repository root for more information.
+
+using System.ComponentModel;
 
 namespace Vortice.Direct2D1.Effects;
 
+/// <summary>
+/// <see cref="https://learn.microsoft.com/en-us/windows/win32/direct2d/convolve-matrix"/>
+/// </summary>
 public sealed class ConvolveMatrix : ID2D1Effect
 {
     public ConvolveMatrix(ID2D1DeviceContext context)
@@ -15,24 +20,27 @@ public sealed class ConvolveMatrix : ID2D1Effect
     {
     }
 
-    public float KernelUnitLehgth
+    [DefaultValue(1.0f)]
+    public float KernelUnitLength
     {
-        set => SetValue((int)ConvolveMatrixProperties.KernelUnitLength, value);
         get => GetFloatValue((int)ConvolveMatrixProperties.KernelUnitLength);
+        set => SetValue((int)ConvolveMatrixProperties.KernelUnitLength, value);
     }
 
     public ConvolveMatrixScaleMode ScaleMode
     {
-        set => SetValue((int)ConvolveMatrixProperties.ScaleMode, value);
         get => GetEnumValue<ConvolveMatrixScaleMode>((int)ConvolveMatrixProperties.ScaleMode);
+        set => SetValue((int)ConvolveMatrixProperties.ScaleMode, value);
     }
 
+    [DefaultValue(3)]
     public int KernelSizeX
     {
         get => (int)GetUIntValue((int)ConvolveMatrixProperties.KernelSizeX);
         set => SetValue((int)ConvolveMatrixProperties.KernelSizeX, (uint)value);
     }
 
+    [DefaultValue(3)]
     public int KernelSizeY
     {
         get => (int)GetUIntValue((int)ConvolveMatrixProperties.KernelSizeY);
@@ -44,20 +52,20 @@ public sealed class ConvolveMatrix : ID2D1Effect
         get
         {
             int size = KernelSizeX * KernelSizeY;
-            float[] value = new float[size];
+            Span<float> value = stackalloc float[size];
             fixed (float* valuePtr = value)
             {
                 GetValue((int)ConvolveMatrixProperties.KernelMatrix, PropertyType.Blob, valuePtr, sizeof(float) * size);
             }
 
-            return value;
+            return value.ToArray();
         }
         set
         {
-            var size = KernelSizeX * KernelSizeY;
+            int size = KernelSizeX * KernelSizeY;
             if (value.Length != size)
             {
-                throw new ArgumentException();
+                throw new ArgumentException($"{nameof(KernelMatrix)} must be KernelSizeX * KernelSizeY");
             }
 
             fixed (void* valuePtr = value)
@@ -67,28 +75,31 @@ public sealed class ConvolveMatrix : ID2D1Effect
         }
     }
 
+    [DefaultValue(1.0f)]
     public float Divisor
     {
-        set => SetValue((int)ConvolveMatrixProperties.Divisor, value);
         get => GetFloatValue((int)ConvolveMatrixProperties.Divisor);
+        set => SetValue((int)ConvolveMatrixProperties.Divisor, value);
     }
 
+    [DefaultValue(0.0f)]
     public float Bias
     {
-        set => SetValue((int)ConvolveMatrixProperties.Bias, value);
         get => GetFloatValue((int)ConvolveMatrixProperties.Bias);
+        set => SetValue((int)ConvolveMatrixProperties.Bias, value);
     }
 
     public Vector2 KernelOffset
     {
-        set => SetValue((int)ConvolveMatrixProperties.KernelOffset, value);
         get => GetVector2Value((int)ConvolveMatrixProperties.KernelOffset);
+        set => SetValue((int)ConvolveMatrixProperties.KernelOffset, value);
     }
 
+    [DefaultValue(false)]
     public bool PreserveAlpha
     {
-        set => SetValue((int)ConvolveMatrixProperties.PreserveAlpha, value);
         get => GetBoolValue((int)ConvolveMatrixProperties.PreserveAlpha);
+        set => SetValue((int)ConvolveMatrixProperties.PreserveAlpha, value);
     }
 
     public BorderMode BorderMode
@@ -97,9 +108,10 @@ public sealed class ConvolveMatrix : ID2D1Effect
         get => GetEnumValue<BorderMode>((int)ConvolveMatrixProperties.BorderMode);
     }
 
+    [DefaultValue(false)]
     public bool ClampOutput
     {
-        set => SetValue((int)ConvolveMatrixProperties.ClampOutput, value);
         get => GetBoolValue((int)ConvolveMatrixProperties.ClampOutput);
+        set => SetValue((int)ConvolveMatrixProperties.ClampOutput, value);
     }
 }