+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Current Stable Release:    3.3.4 (October 2023)
 -----------------------------------------------
 Release:     3.4.X (March 2024)
 
-- [ENH] 
+- [ENH] Update DirectX Headers to 1.611.1
+- [ENH] Update DirectStorage to 1.2.1
 
 -----------------------------------------------
 Release:     3.3.4 (October 2023)